% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_network.R
\name{Cons_net}
\alias{Cons_net}
\title{Construct Network for plotting}
\usage{
Cons_net(mouse = FALSE)
}
\arguments{
\item{mouse}{A logical indicating whether to construct the network for mouse or human. Default is FALSE for human.}
}
\value{
net: TF-Target network

The constructed network to a file named "TF_Network.txt"
}
\description{
It constructs the TF-Target network and returns network file ("TF_Network.txt") if you want to plot your network using other software like Cytoscape.
}
