% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_analysis.R
\name{DiffDESeq2}
\alias{DiffDESeq2}
\title{DEG Analysis of RNA-seq Data using DESeq with P value}
\usage{
DiffDESeq2(counts, phenodata, complist, logfc = 0, pval = 0.05, padj = TRUE)
}
\arguments{
\item{counts}{A count matrix where rows are genes and columns are samples.}

\item{phenodata}{A data frame containing the phenotype data for each sample.}

\item{complist}{A character vector containing the names of the samples to compare.}

\item{logfc}{The log2 fold change cutoff for differential expression. Default is 0.}

\item{pval}{The adjusted p-value or p-value cutoff for differential expression. Default is 0.05.}

\item{padj}{Logical indicating whether to use adjusted p-values or not. Default is TRUE.}
}
\value{
DEresult: A list containing the following items:
table: A data frame with the results of the differential expression analysis.
rank_vector: A vector containing the rank of each gene based on the effect size.
degs: A character vector containing the names of the differentially expressed genes.
e: A matrix of normalized expression values.
up: A character vector containing the names of the upregulated genes.
down: A character vector containing the names of the downregulated genes.
nosig: A character vector containing the names of the non-differentially expressed genes.
}
\description{
This function performs differential expression analysis using the DESeq2 package.
}
