% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_analysis.R
\name{Difflimma}
\alias{Difflimma}
\title{DEG Analysis of RNA-seq Data using limma + Voom with P value}
\usage{
Difflimma(counts, phenodata, complist, logfc = 1, pval = 0.05, padj = TRUE)
}
\arguments{
\item{counts}{Processed gene expression count data}

\item{phenodata}{pData that provides batch & experimental conditions}

\item{complist}{a vector of multiple comparisons in the format of contrasts in limma (e.g. c("A-B", "A-C", "B-C"))}

\item{logfc}{(optional) log fold change constraints for DEGs}

\item{pval}{P-value cutoff for DEG analysis (default: 0.05)}

\item{padj}{Use adjusted p-value or not (default: TRUE)}
}
\value{
DEresult: a list of DEG results, including those for each single comparison and those for the overall comparison.
Each DEG result is in the format of A list containing:
table: table of DEG results.
rank_vector: a vector of t-statistics for every gene.
degs: a vector of gene symbols for DEGs.
e: expression data (CPM).
e_batch: batch corrected expression.
}
\description{
DEG Analysis of RNA-seq Data using limma + Voom with P value
}
