% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_network.R
\name{plot_3d_tfnetwork}
\alias{plot_3d_tfnetwork}
\title{Plot a 3D TF network using tkplot from igraph}
\usage{
plot_3d_tfnetwork(network)
}
\arguments{
\item{network}{an igraph object representing the network to plot}
}
\value{
A 3D network graph plot.
}
\description{
This function plots the input network as a 3D network graph using the rgl package.
}
\examples{
plot_3d_tfnetwork(network)
}
