% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_network.R
\name{plot_interactive_tfnetwork}
\alias{plot_interactive_tfnetwork}
\title{Plot interactive TF network}
\usage{
plot_interactive_tfnetwork(
  network,
  canvas.width = 500,
  canvas.height = 500,
  vertex.label.family = "sans"
)
}
\arguments{
\item{network}{TF network}
}
\value{
interactive TF network plot
}
\description{
Plot interactive TF network
}
