% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_network.R
\name{plot_tfnetwork}
\alias{plot_tfnetwork}
\title{Plot TF-Target network}
\usage{
plot_tfnetwork(
  net,
  nodeshape = "circle",
  TF_color = "#F0E442",
  Up_color = "#fb8072",
  Down_color = "#80b1d3",
  TF_alpha = 0.75,
  Up_alpha = 0.3,
  Down_alpha = 0.3,
  TF_size = 10,
  Target_size = 8,
  TF_labelsize = 0.7,
  Target_labelsize = 0.7,
  label_color = "black",
  edge_color = "grey",
  edge_alpha = 0.75,
  netlayout = NULL,
  pdf.name = "TFnet.pdf",
  pdf.height = 15,
  pdf.width = 15,
  edge.arrow.size = 0.4,
  vertex.label.family = "sans"
)
}
\arguments{
\item{nodeshape}{shape of the nodes}
}
\value{
TFnet.pdf: network plot
}
\description{
Plot TF-Target network
}
