% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_analysis.R
\name{pre}
\alias{pre}
\title{Preprocess expression coutns}
\usage{
pre(
  counts,
  groups,
  mouse = FALSE,
  min.count = 1,
  min.total.count = 2,
  large.n = 1,
  min.prop = 0.1
)
}
\arguments{
\item{counts}{A matrix or data frame containing the raw count data.}

\item{groups}{A factor vector indicating the group membership of each sample.}

\item{mouse}{A logical indicating whether the input data is from mouse or human.}

\item{min.count}{The minimum number of counts for a gene to be considered expressed in a sample.}

\item{min.total.count}{The minimum total count across all samples for a gene to be considered expressed.}

\item{large.n}{The threshold number of samples in which a gene must be expressed to avoid filtering due to low expression.}

\item{min.prop}{The minimum proportion of samples in which a gene must be expressed to avoid filtering due to low expression.}
}
\value{
x$counts: processed count matrix
}
\description{
Preprocess expression coutns. The code is primarily based on the NetAct package, which can be found at https://github.com/lusystemsbio/NetAct.
}
