% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetAct.R
\name{Reselect_TFs_p}
\alias{Reselect_TFs_p}
\title{TF reselection with P-value}
\usage{
Reselect_TFs_p(GSEArslt, pval = 0.05, combine_TFs = TRUE, ntop = NULL)
}
\arguments{
\item{GSEArslt}{GSEA results from TF_Selection}

\item{pval}{p-value cutoff (default: 0.05)}

\item{combine_TFs}{whether combine selected TFs from multiple comparisons or not (default: TRUE)}

\item{ntop}{the number of top genes (selection by the top genes) (default: NULL, no selection by the top genes)}
}
\value{
tfs: a vector of selected TFs
}
\description{
Reselecting TFs using gene set enrichement analysis (GSEA) using an adjusted set of parameters (work together with TF_Selection). The code is primarily based on the NetAct package, which can be found at https://github.com/lusystemsbio/NetAct.
}
