% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetAct.R
\name{row_norm}
\alias{row_norm}
\title{Row normalization (standardization)}
\usage{
row_norm(data)
}
\arguments{
\item{data}{gene expression matrix}
}
\value{
norm_data: standardized gene expression matrix
}
\description{
Row normalization of gene expression matrix. The code is primarily based on the NetAct package, which can be found at https://github.com/lusystemsbio/NetAct.
}
