% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetAct.R
\name{TF_heatmap}
\alias{TF_heatmap}
\title{Plotting TF gene expresion & activity heatmap}
\usage{
TF_heatmap(
  new_activity,
  eset,
  activity_range = c(-1, 0, 1),
  activity_color = c("#4090BD", "#F5F5F5", "#E85355"),
  exp_range = c(-2, 0, 2),
  exp_color = c("#208421", "#f0e9db50", "#b40001")
)
}
\arguments{
\item{new_activity}{Matrix. TF activity matrix}

\item{eset}{ExpressionSet of gene expression data}

\item{activity_range}{Range of activity in heatmap}

\item{activity_color}{Colors of activity in heatmap}

\item{exp_range}{Range of gene expression Z-score in heatmap}

\item{exp_color}{Colors of gene expression Z-score in heatmap}
}
\value{
Heatmap plotting object
}
\description{
Plotting TF gene expresion & activity heatmap
}
