% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_analysis.R
\name{Cons_exp}
\alias{Cons_exp}
\title{Construct expression matrix}
\usage{
Cons_exp(files, group, samplenames, compList)
}
\arguments{
\item{files}{pre-defined inputs}
}
\value{
x: expression matrix for later analysis
}
\description{
Construct expression matrix
}
