% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_network.R
\name{Cons_net}
\alias{Cons_net}
\title{Contruct Network for plotting}
\usage{
Cons_net(mouse = FALSE)
}
\arguments{
\item{mouse}{use mouse genome or not}
}
\value{
net: TF-Target network
}
\description{
Contruct Network for plotting
}
