% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_analysis.R
\name{Difflimma}
\alias{Difflimma}
\title{DEG Analysis of RNA-seq Data using limma + Voom with P value}
\usage{
Difflimma(counts, phenodata, complist, logfc = 1, pval = 0.05, padj = TRUE)
}
\arguments{
\item{counts}{A matrix of raw counts.}

\item{phenodata}{A data frame containing phenotype information.}

\item{complist}{A list of comparisons to be performed.}

\item{logfc}{A numeric value indicating the log fold change threshold (default: 1).}

\item{pval}{A numeric value indicating the p-value threshold (default: 0.05).}

\item{padj}{A logical value indicating whether to adjust p-values (default: TRUE).}
}
\value{
A list containing the following elements:
\item{table}{The differential expression table.}
\item{rank_vector}{A rank vector.}
\item{degs}{A list of differentially expressed genes.}
\item{e}{Normalized expression values.}
}
\description{
This function performs differential expression analysis using the limma package. The code is primarily based on the NetAct package, which can be found at https://github.com/lusystemsbio/NetAct.
}
