% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_network.R
\name{plot_3d_tfnetwork}
\alias{plot_3d_tfnetwork}
\title{Plot 3D TF network}
\usage{
plot_3d_tfnetwork(network)
}
\arguments{
\item{network}{TF network}
}
\value{
3d TF network plot
}
\description{
Plot 3D TF network
}
