% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_network.R
\name{plot_highlighted_tfnetwork}
\alias{plot_highlighted_tfnetwork}
\title{Plot highlighted TF network}
\usage{
plot_highlighted_tfnetwork(
  network,
  highlight_node,
  highlight_TF_color = "#f19d00",
  highlight_edge_color = "#e41a1c",
  netlayout = NULL,
  pdf_height = 15,
  pdf_width = 15,
  vertex_label_family = "sans"
)
}
\arguments{
\item{highlight_node}{The key node you want to highlight}
}
\value{
TF_highlight.pdf: highlighted network plot
}
\description{
Plot highlighted TF network
}
