% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_network.R
\name{plot_interactive_tfnetwork}
\alias{plot_interactive_tfnetwork}
\title{Plot an interactive TF network using tkplot from igraph}
\usage{
plot_interactive_tfnetwork(
  network,
  canvas.width = 500,
  canvas.height = 500,
  vertex.label.family = "sans"
)
}
\arguments{
\item{network}{The igraph object representing the TF network to be plotted}

\item{canvas.width}{The width of the canvas for the tkplot window}

\item{canvas.height}{The height of the canvas for the tkplot window}

\item{vertex.label.family}{The font family to use for vertex labels}
}
\value{
interactive TF network plot
}
\description{
This function plots an interactive transcription factor (TF) network using the tkplot function from the igraph package. The resulting plot can be manipulated using mouse gestures and keyboard commands.
}
