% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEG_analysis.R
\name{pre}
\alias{pre}
\title{Preprocess expression coutns}
\usage{
pre(
  counts,
  groups,
  mouse = FALSE,
  min.count = 1,
  min.total.count = 2,
  large.n = 1,
  min.prop = 0.1
)
}
\arguments{
\item{mouse}{use mouse genome or not}
}
\value{
x$counts: processed count matrix
}
\description{
Preprocess expression coutns
}
