% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetAct.R
\name{TF_heatmap}
\alias{TF_heatmap}
\title{TF reselection with P-value}
\usage{
TF_heatmap(
  new_activity,
  eset,
  activity_range = c(-1, 0, 1),
  activity_color = c("#4090BD", "#F5F5F5", "#E85355"),
  exp_range = c(-2, 0, 2),
  exp_color = c("#208421", "#f0e9db50", "#b40001")
)
}
\arguments{
\item{new_activity}{a numeric matrix or data.frame containing the activity values of transcription factors (TFs)}

\item{eset}{an ExpressionSet or numeric matrix or data.frame containing the expression values of genes}

\item{activity_range}{a vector with three elements specifying the range of the activity values of TFs for color scaling}

\item{activity_color}{a vector with three colors specifying the color scheme for the activity heatmap}

\item{exp_range}{a vector with three elements specifying the range of the expression values of genes for color scaling}

\item{exp_color}{a vector with three colors specifying the color scheme for the expression heatmap}
}
\value{
A heatmap object generated from the given TF activity and gene expression data
}
\description{
Plotting heatmap of TF activity & TF gene expresion. The code is primarily based on the NetAct package, which can be found at https://github.com/lusystemsbio/NetAct.
}
